Adicionando funcionalidades de detalhes e edicao de visitantes

+++ cadastro_visitantes/cadastro_app/models.py
@@ -20,6 +20,6 @@ class Visitante(models.Model):
     ]
     tipo = models.CharField(max_length=10, choices=tipo_visitante)
     nome = models.TextField(max_length=255)
-    rg = models.FloatField()
+    rg = models.IntegerField()
     motivo = models.TextField(max_length=255)
     data = models.DateField(default=datetime.now)
